Overhaul Chromatik class hierarchy again to provide Property accessors
For asynchronous chromasthetiation the relevant parameters need to be
"frozen" to their state at request submission. This is achieved
through (deep) copies of the objects holding this state.
Unfortunately instances of "SimpleProperty*" don't implement the
Cloneable interface, so there needed to be some kind of abstraction
between objects relying on "Property" to manage application
preferences and those holding the frozen state for asynchronous
chromasthetiation.